software/tester/requirements.txt
ntplib
pyserial
